clean code: python serializer
